+++ core/src/cli.js
@@ -30,7 +30,7 @@ const Terracotta = require('./terracotta');
 program
     .name('wow')
     .description('Minecraft 服务器管理工具 - 默认优先，可以修改')
-    .version('3.3.12', '-V');
+    .version('3.3.13', '-V');
 
 // ==================== init ====================